refactor: conditionally import custom component classes
