+++ src/async-utils/platform.ts
