+++ src/tailwindcss/custom-component-classes.async.v4.ts
@@ -3,8 +3,6 @@ import { readFileSync } from "node:fs";
 import { fork } from "@eslint/css-tree";
 import { tailwind4 } from "tailwind-csstree";
 
-import { invalidateByModifiedDate, withCache } from "../utils/cache.js";
-
 import type { CssNode } from "@eslint/css-tree";
 
 import type {
@@ -15,11 +13,11 @@ import type {
 
 const { findAll, generate, parse } = fork(tailwind4);
 
-export const getCustomComponentClasses = ({ configPath }: GetCustomComponentClassesRequest) => withCache<GetCustomComponentClassesResponse>("ast", () => {
+export function getCustomComponentClasses({ configPath }: GetCustomComponentClassesRequest): GetCustomComponentClassesResponse {
   const entryFile = readFileSync(configPath, "utf-8");
   const ast = parse(entryFile);
   return getCustomComponentUtilities(ast);
-}, date => invalidateByModifiedDate(date, configPath));
+}
 
 function getCustomComponentUtilities(ast: CssNode) {
   const customComponentUtilities: string[] = [];
