+++ src/tailwindcss/custom-component-classes.ts
@@ -1,14 +1,15 @@
 // runner.js
 
 
+import { resolve } from "node:path";
+
+import { createSyncFn } from "synckit";
+
 import { getTailwindConfigPath } from "better-tailwindcss:tailwindcss/config.js";
-import {
-  getCustomComponentClasses as getCustomComponentClassesV3
-} from "better-tailwindcss:tailwindcss/custom-component-classes.v3.js";
-import {
-  getCustomComponentClasses as getCustomComponentClassesV4
-} from "better-tailwindcss:tailwindcss/custom-component-classes.v4.js";
-import { getTailwindcssVersion, TailwindcssVersion } from "better-tailwindcss:utils/version.js";
+import { invalidateByModifiedDate, withCache } from "better-tailwindcss:utils/cache.js";
+import { getWorkerOptions } from "better-tailwindcss:utils/worker.js";
+
+import type { Async } from "better-tailwindcss:types/async.js";
 
 
 export interface GetCustomComponentClassesRequest {
@@ -18,12 +19,23 @@ export interface GetCustomComponentClassesRequest {
 export type GetCustomComponentClassesResponse = string[];
 
 export function getCustomComponentClasses({ configPath, cwd }: { configPath: string | undefined; cwd: string; }) {
-  const version = getTailwindcssVersion();
   const { path } = getTailwindConfigPath({ configPath, cwd });
 
-  const customComponentClasses = version.major <= TailwindcssVersion.V3
-    ? getCustomComponentClassesV3({ configPath: path })
-    : getCustomComponentClassesV4({ configPath: path });
+  return withCache("custom-components", () => {
+    return getCustomComponentClassesSync({ configPath: path });
+  }, date => invalidateByModifiedDate(date, path));
+}
 
-  return customComponentClasses;
+const getCustomComponentClassesSync = createSyncFn<
+  Async<GetCustomComponentClassesRequest, GetCustomComponentClassesResponse>
+>(getWorkerPath(), getWorkerOptions());
+
+function getWorkerPath() {
+  return resolve(getCurrentDirectory(), "./custom-component-classes.async.js");
+}
+
+function getCurrentDirectory() {
+  // eslint-disable-next-line eslint-plugin-typescript/prefer-ts-expect-error
+  // @ts-ignore - `import.meta` doesn't exist in CommonJS -> will be transformed in build step
+  return import.meta.dirname;
 }
